add new Archive page + refactored SoldItems component
